+++ NEWS.md
@@ -1,7 +1,8 @@
+- Add plot for confidence intervals (`mlr3inferr`)
+
 # mlr3viz (development version)
 
 - Add plot for `LearnerSurvCoxPH`.
-- Add plot for confidence intervals (`mlr3inferr`)
 
 # mlr3viz 0.9.0